requirements.txt
streamlit>=1.30.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
langgraph>=0.2.0
langchain-core>=0.2.0
python-dotenv>=1.0.0
arxiv>=2.1.0
wikipedia>=1.4.0
tavily-python>=0.3.0